+++ src/framework/SexyAppBase.h
@@ -106,7 +106,7 @@ using SharedImageMap = std::map<std::string, std::unique_ptr<Image>>;
 
 class SexyAppBase : public ButtonListener, public DialogListener {
 public:
-    ulong mRandSeed;
+    uint32_t mRandSeed;
 
     std::string mCompanyName;
     std::string mFullCompanyName;
@@ -449,10 +449,10 @@ public:
     void MirrorImage(Image *theImage);
     void FlipImage(Image *theImage);
     //	void					RotateImageHue(Sexy::MemoryImage *theImage, int theDelta);
-    ulong HSLToRGB(int h, int s, int l);
-    ulong RGBToHSL(int r, int g, int b);
-    void HSLToRGB(const ulong *theSource, ulong *theDest, int theSize);
-    void RGBToHSL(const ulong *theSource, ulong *theDest, int theSize);
+    uint32_t HSLToRGB(int h, int s, int l);
+    uint32_t RGBToHSL(int r, int g, int b);
+    void HSLToRGB(const uint32_t *theSource, uint32_t *theDest, int theSize);
+    void RGBToHSL(const uint32_t *theSource, uint32_t *theDest, int theSize);
 
     //	void					AddMemoryImage(MemoryImage* theMemoryImage);
     //	void					RemoveMemoryImage(MemoryImage* theMemoryImage);
@@ -546,7 +546,7 @@ public:
     bool RegistryWriteString(const std::string &theValueName, const std::string &theString);
     bool RegistryWriteInteger(const std::string &theValueName, int theValue);
     bool RegistryWriteBoolean(const std::string &theValueName, bool theValue);
-    bool RegistryWriteData(const std::string &theValueName, const uint8_t *theValue, ulong theLength);
+    bool RegistryWriteData(const std::string &theValueName, const uint8_t *theValue, uint32_t theLength);
     bool RegistryEraseKey(const SexyString &theKeyName);
     static void RegistryEraseValue(const SexyString &theValueName);
