+++ src/framework/graphics/Graphics.cpp
@@ -1053,7 +1053,7 @@ int Graphics::WriteWordWrapped(
     if (theLineSpacing == -1) theLineSpacing = aFont->GetLineSpacing();
 
     SexyString aCurString;
-    ulong aCurPos = 0;
+    uint32_t aCurPos = 0;
     int aLineStartPos = 0;
     int aCurWidth = 0;
     SexyChar aCurChar = 0;
