+++ src/framework/misc/Buffer.h
@@ -52,7 +52,7 @@ public:
     const uint8_t *GetDataPtr() const;
     int GetDataLen() const;
     int GetDataLenBits() const;
-    ulong GetCRC32(ulong theSeed = 0) const;
+    uint32_t GetCRC32(uint32_t theSeed = 0) const;
 
     bool AtEnd() const;
     bool PastEnd() const;
